+++ src/main/java/com/cMall/feedShop/common/aop/ResponseFormatAspect.java
@@ -17,21 +17,22 @@ public class ResponseFormatAspect {
         try {
             Object result = joinPoint.proceed();
 
-            // 이미 ApiResponse 타입이면 그대로 반환
-            if (result instanceof ApiResponse) {
-                return result;
+            if (result instanceof ApiResponse<?> apiResponse) {
+                String customMessage = apiResponseFormat.message();
+                if (customMessage != null && !customMessage.isEmpty()) {
+                    apiResponse.setMessage(customMessage);
+                }
+                return apiResponse;
+            } else {
+                String messageToUse = apiResponseFormat.message().isEmpty() ?
+                        "요청이 성공했습니다." : apiResponseFormat.message();
+                return ApiResponse.success(messageToUse, result);
             }
-
-            // 성공 응답으로 래핑
-            String message = apiResponseFormat.message().isEmpty() ?
-                    "요청이 성공했습니다." : apiResponseFormat.message();
-
-            return ApiResponse.success(message, result);
 
         } catch (Exception e) {
             log.error("API 실행 중 오류 발생 - Method: {}, Error: {}",
                     joinPoint.getSignature().getName(), e.getMessage());
-            throw e; // 예외는 GlobalExceptionHandler에서 처리
+            throw e;
         }
     }
 }
