+++ src/main/java/com/cMall/feedShop/common/dto/ApiResponse.java
@@ -1,13 +1,11 @@
 package com.cMall.feedShop.common.dto;
 
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
+import lombok.*;
 
 import java.time.LocalDateTime;
 
 @Getter
+@Setter
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
