+++ src/main/java/com/cMall/feedShop/config/BaseSecurityConfig.java
@@ -1,0 +1,82 @@
+//package com.cMall.feedShop.config;
+//
+//import com.cMall.feedShop.user.infrastructure.security.JwtAuthenticationFilter;
+//import lombok.RequiredArgsConstructor;
+//import org.springframework.context.annotation.Bean;
+//import org.springframework.context.annotation.Configuration;
+//import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity;
+//import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+//import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+//import org.springframework.security.config.http.SessionCreationPolicy;
+//import org.springframework.security.web.SecurityFilterChain;
+//import org.springframework.security.web.authentication.HttpStatusEntryPoint;
+//import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
+//import org.springframework.web.cors.CorsConfiguration;
+//import org.springframework.web.cors.CorsConfigurationSource;
+//import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
+//import org.springframework.http.HttpStatus;
+//
+//import java.util.Arrays;
+//import java.util.List;
+//
+//@Configuration
+//@EnableWebSecurity
+//@EnableMethodSecurity
+//@RequiredArgsConstructor
+//public class BaseSecurityConfig {
+//
+//    private final JwtAuthenticationFilter jwtAuthenticationFilter;
+//
+//    @Bean
+//    public SecurityFilterChain securityFilterChain(HttpSecurity httpSecurity) throws Exception {
+//        httpSecurity
+//                .csrf(csrf -> csrf.disable())
+//                .cors(cors -> cors.configurationSource(corsConfigurationSource()))
+//
+//                .sessionManagement(session -> session.sessionCreationPolicy(SessionCreationPolicy.STATELESS))
+//
+//                .exceptionHandling(exceptionHandling -> exceptionHandling
+//                        .authenticationEntryPoint(new HttpStatusEntryPoint(HttpStatus.UNAUTHORIZED))
+//                )
+//
+//                .authorizeHttpRequests(auth -> auth
+//                    .requestMatchers(
+//                      "/api/auth/login",
+//                      "/api/auth/signup",
+//                      "/api/auth/verify-email",
+//                      "/public/**",
+//                      "/swagger-ui/**",
+//                      "/v3/api-docs/**",
+//                      "/swagger-resources/**",
+//                      "/api/products",
+//                      "/api/products/**",
+//                      "/api/events", "/api/events/**",
+//                      "/api/reviews/products/**", "/api/reviews/{reviewId}"
+//                    ).permitAll()
+//                    .requestMatchers("/api/users/admin/**").hasRole("ADMIN")
+//                    .requestMatchers("/api/seller/**").hasRole("SELLER")
+//                    .requestMatchers("/actuator/**").hasRole("ADMIN")
+//                    .anyRequest().authenticated()
+//                )
+//                .formLogin(formLogin -> formLogin.disable())
+//                .httpBasic(httpBasic -> httpBasic.disable())
+//                .addFilterBefore(jwtAuthenticationFilter, UsernamePasswordAuthenticationFilter.class);
+//
+//        return httpSecurity.build();
+//    }
+//
+//    @Bean
+//    public CorsConfigurationSource corsConfigurationSource() {
+//        CorsConfiguration config = new CorsConfiguration();
+//        config.setAllowedOrigins(Arrays.asList(
+//                "https://feedshop-frontend.vercel.app/",
+//                "http://localhost:3000"
+//        ));
+//        config.setAllowedMethods(List.of("GET", "POST", "PUT", "DELETE", "OPTIONS"));
+//        config.setAllowedHeaders(List.of("*"));
+//        config.setAllowCredentials(true);
+//        UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
+//        source.registerCorsConfiguration("/**", config);
+//        return source;
+//    }
+//}
