+++ src/main/java/com/cMall/feedShop/user/domain/exception/DuplicateEmailException.java
@@ -1,0 +1,10 @@
+package com.cMall.feedShop.user.domain.exception;
+
+import com.cMall.feedShop.common.exception.BusinessException;
+import com.cMall.feedShop.common.exception.ErrorCode;
+
+public class DuplicateEmailException extends BusinessException {
+    public DuplicateEmailException() {
+        super(ErrorCode.DUPLICATE_EMAIL);
+    }
+}
