+++ src/main/java/com/cMall/feedShop/user/domain/exception/UserNotFoundException.java
@@ -1,0 +1,14 @@
+package com.cMall.feedShop.user.domain.exception;
+
+import com.cMall.feedShop.common.exception.BusinessException;
+import com.cMall.feedShop.common.exception.ErrorCode;
+
+public class UserNotFoundException extends BusinessException {
+    public UserNotFoundException() {
+        super(ErrorCode.USER_NOT_FOUND);
+    }
+
+    public UserNotFoundException(String message) {
+        super(ErrorCode.USER_NOT_FOUND, message);
+    }
+}
